refactored some files due to feedback
