+++ Src/Test/SDK/WorkflowClient/Temporal.Sdk.WorkflowClient.Test.E2EInt/public/SimpleClientInvocationsE2ETest.cs
@@ -4,7 +4,7 @@ using Temporal.TestUtil;
 namespace Temporal.Sdk.WorkflowClient.Test.E2EInt
 {
     // ReSharper disable once UnusedType.Global
-    public class SimpleClientInvocationsE2ETest : AbstractSimpleClientInvocationsE2ETest
+    public class SimpleClientInvocationsE2ETest : SimpleClientInvocationsE2ETestBase
     {
         public SimpleClientInvocationsE2ETest(ITestOutputHelper cout)
             : base(cout, TestTlsOptions.None, 7233)
