+++ Src/Test/SDK/WorkflowClient/Temporal.Sdk.WorkflowClient.Test.E2EInt/public/SimpleClientInvocationsE2EWithTlsTest.cs
@@ -4,7 +4,7 @@ using Xunit.Abstractions;
 namespace Temporal.Sdk.WorkflowClient.Test.E2EInt
 {
     // ReSharper disable once UnusedType.Global
-    public class SimpleClientInvocationsE2EWithTlsTest : AbstractSimpleClientInvocationsE2ETest
+    public class SimpleClientInvocationsE2EWithTlsTest : SimpleClientInvocationsE2ETestBase
     {
         public SimpleClientInvocationsE2EWithTlsTest(ITestOutputHelper cout)
             : base(cout, TestTlsOptions.Server, 7234)
