+++ Src/Test/SDK/WorkflowClient/Temporal.Sdk.WorkflowClient.Test.Int/public/TemporalClientMutualTlsTest.cs
@@ -4,7 +4,7 @@ using Xunit.Abstractions;
 namespace Temporal.Sdk.WorkflowClient.Test.Int
 {
     // ReSharper disable once UnusedType.Global
-    public class TemporalClientMutualTlsTest : AbstractTemporalClientTest
+    public class TemporalClientMutualTlsTest : TemporalClientTestBase
     {
         public TemporalClientMutualTlsTest(ITestOutputHelper cout)
             : base(cout, TestTlsOptions.Mutual, 7234)
