+++ Src/Test/SDK/WorkflowClient/Temporal.Sdk.WorkflowClient.Test.Int/public/TemporalClientTest.cs
@@ -4,7 +4,7 @@ using Temporal.TestUtil;
 namespace Temporal.Sdk.WorkflowClient.Test.Int
 {
     // ReSharper disable once UnusedType.Global
-    public class TemporalClientTest : AbstractTemporalClientTest
+    public class TemporalClientTest : TemporalClientTestBase
     {
         public TemporalClientTest(ITestOutputHelper cout)
             : base(cout, TestTlsOptions.None)
