+++ Src/Test/SDK/WorkflowClient/Temporal.Sdk.WorkflowClient.Test.Int/public/TemporalClientTlsTest.cs
@@ -4,7 +4,7 @@ using Xunit.Abstractions;
 namespace Temporal.Sdk.WorkflowClient.Test.Int
 {
     // ReSharper disable once UnusedType.Global
-    public class TemporalClientTlsTest : AbstractTemporalClientTest
+    public class TemporalClientTlsTest : TemporalClientTestBase
     {
         public TemporalClientTlsTest(ITestOutputHelper cout)
             : base(cout, TestTlsOptions.Server, 7235)
